+++ rio_tiler/io/stac.py
@@ -2,7 +2,7 @@
 
 import functools
 import json
-from typing import Dict, Iterator, Optional, Set, Type, Union
+from typing import Any, Dict, Iterator, Optional, Set, Type, Union
 from urllib.parse import urlparse
 
 import attr
@@ -31,13 +31,14 @@ DEFAULT_VALID_TYPE = {
 
 
 @functools.lru_cache(maxsize=512)
-def fetch(filepath: str) -> Dict:
+def fetch(filepath: str, **kwargs: Any) -> Dict:
     """Fetch STAC items.
 
     A LRU cache is set on top of this function.
 
     Args:
         filepath (str): STAC item URL.
+        kwargs (any): additional options to pass to client.
 
     Returns:
         dict: STAC Item content.
@@ -47,10 +48,10 @@ def fetch(filepath: str) -> Dict:
     if parsed.scheme == "s3":
         bucket = parsed.netloc
         key = parsed.path.strip("/")
-        return json.loads(aws_get_object(bucket, key))
+        return json.loads(aws_get_object(bucket, key, **kwargs))
 
     elif parsed.scheme in ["https", "http", "ftp"]:
-        return requests.get(filepath).json()
+        return requests.get(filepath, **kwargs).json()
 
     else:
         with open(filepath, "r") as f:
@@ -130,7 +131,8 @@ class STACReader(MultiBaseReader):
         include_asset_types (set of string, optional): Only include some assets base on their type.
         exclude_asset_types (set of string, optional): Exclude some assets base on their type.
         reader (rio_tiler.io.BaseReader, optional): rio-tiler Reader. Defaults to `rio_tiler.io.COGReader`.
-        reader_options (dict, optional): additional option to forward to the Reader. Defaults to `{}`.
+        reader_options (dict, optional): Additional option to forward to the Reader. Defaults to `{}`.
+        fetch_options (dict, optional): Options to pass to `rio_tiler.io.stac.fetch` function fetching the STAC Items. Defaults to `{}`.
 
     Examples:
         >>> with STACReader(stac_path) as stac:
@@ -162,11 +164,12 @@ class STACReader(MultiBaseReader):
     exclude_asset_types: Optional[Set[str]] = attr.ib(default=None)
     reader: Type[BaseReader] = attr.ib(default=COGReader)
     reader_options: Dict = attr.ib(factory=dict)
+    fetch_options: Dict = attr.ib(factory=dict)
 
     def __attrs_post_init__(self):
         """Fetch STAC Item and get list of valid assets."""
         self.item = self.item or pystac.Item.from_dict(
-            fetch(self.filepath), self.filepath
+            fetch(self.filepath, **self.fetch_options), self.filepath
         )
         self.bounds = self.item.bbox
         self.assets = list(