+++ CHANGES.md
@@ -1,8 +1,13 @@
 ## Unreleased
 
 * add support for setting the S3 endpoint url via the `AWS_S3_ENDPOINT` environment variables in `aws_get_object` function using boto3 (https://github.com/cogeotiff/rio-tiler/pull/394)
-
 * make `ImageStatistics.valid_percent` a value between 0 and 100 (instead of 0 and 1) (author @param-thakker, https://github.com/cogeotiff/rio-tiler/pull/400)
+* add `fetch_options` to `STACReader` to allow custom configuration to the fetch client (https://github.com/cogeotiff/rio-tiler/pull/404)
+
+    ```python
+    with STACReader("s3://...", fetch_options={"request_pays": True}):
+        pass
+    ```
 
 ## 2.1.0 (2021-05-17)